requirements.txt
pandas>=2.0.3
numpy>=1.24.3
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
huggingface-hub>=0.20.3
langchain>=0.1.0
python-dotenv>=1.0.0
